--- src/components/About/about.d.ts
@@ -1,7 +1,0 @@
-/*
-* интерфейс объекта ответа, возвращаемого микросервисом "informator"
-*/
-interface IAbout {
-  alias: string
-  mdInfo: string
-}
