--- src/components/Popup/popup.d.ts
@@ -1,1 +1,0 @@
-type popupMode = "success" | "danger" | undefined
