--- src/components/Setting/setting.d.ts
@@ -1,12 +1,0 @@
-interface IRole extends IRow {
-  directings: IDirecting[]
-}
-
-interface IDirecting extends IRow {
-  tasks: ITask[]
-}
-
-interface ITask extends IRow {
-  actions: IRow[]
-}
-
