--- src/components/User/user.d.ts
@@ -1,8 +1,0 @@
-interface IUser {
-  email: string
-  rank: string
-  roles: IRole[]
-  name?: string
-  photo?: string
-  status?: string
-}
