--- src/components/catalog/Test/Test2.tsx
@@ -1,41 +1,0 @@
-export default function F(){return <></>}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-// import { useState } from "react";
-
-// export default function Test2(){
-//   return <>
-//     <Foo />
-//     <Foo />
-//   </>
-// }
-
-// function Foo(){
-//   const [x, setX] = useState(0)
-//   return <p onClick={() => setX(x+1)}>click me {x}</p>
-// }
