--- src/contexts/ThemeProvider.d.ts
@@ -1,6 +1,0 @@
-type StyleTheme = "dark" | "light";
-
-interface ITheme {
-    theme: StyleTheme;
-    setTheme: Dispatch<SetStateAction<StyleTheme>>;
-}
