+++ src/types/global.d.ts
@@ -2,5 +2,3 @@ declare module "*.module.scss" {
   const classes: { readonly [key: string]: string };
   export default classes;
 }
-
-type ServiceName = "bridge" | "mauth" | "informator";
